refactor(forms): isolate multiselect terminal editor
