+++ packages/bijou/src/core/forms/multiselect.ts
@@ -4,16 +4,9 @@ import { resolveCtx } from '../resolve-ctx.js';
 import {
   formatFormTitle,
   renderNumberedOptions,
-  terminalRenderer,
   formDispatch,
-  createStyledFn,
-  createBoldFn,
-  clampScroll,
-  handleVerticalNav,
-  isKey,
-  subscribeFormKeyInput,
 } from './form-utils.js';
-import { sanitizePositiveInt } from '../numeric.js';
+import { interactiveMultiselect } from './multiselect-interactive.js';
 
 /**
  * Options for the multi-select field.
@@ -68,106 +61,4 @@ async function numberedMultiselect<T>(options: MultiselectOptions<T>, ctx: Bijou
     .map((s) => parseInt(s.trim(), 10) - 1)
     .filter((i) => i >= 0 && i < options.options.length);
   return indices.map((i) => optionAt(options.options, i).value);
-}
-
-/**
- * Display a keyboard-navigable multi-select menu using raw terminal input.
- *
- * Supports arrow keys and j/k for navigation, Space to toggle selection,
- * Enter to confirm, Ctrl+C or Escape to cancel (returns empty array).
- *
- * @param options - Multiselect field configuration.
- * @param ctx - Bijou context.
- * @returns Array of selected option values in index order.
- */
-async function interactiveMultiselect<T>(options: MultiselectOptions<T>, ctx: BijouContext): Promise<T[]> {
-  const noColor = ctx.theme.noColor;
-  const styledFn = createStyledFn(ctx);
-  const boldFn = createBoldFn(ctx);
-  const term = terminalRenderer(ctx);
-  const maxVisible = sanitizePositiveInt(options.maxVisible, 7);
-
-  let cursor = 0;
-  let scrollOffset = 0;
-  const selected = new Set<number>();
-
-  if (options.defaultValues !== undefined) {
-    for (let i = 0; i < options.options.length; i++) {
-      if (options.defaultValues.some((dv) => Object.is(dv, optionAt(options.options, i).value))) {
-        selected.add(i);
-      }
-    }
-  }
-
-  function visibleOptions(): SelectOption<T>[] {
-    return options.options.slice(scrollOffset, scrollOffset + maxVisible);
-  }
-
-  function renderLineCount(): number {
-    return 1 + Math.min(options.options.length, maxVisible);
-  }
-
-  function render(): void {
-    const label = formatFormTitle(options.title, ctx);
-    term.hideCursor();
-    const hint = styledFn(ctx.semantic('muted'), '(space to toggle, enter to confirm)');
-    term.writeLine(`${label}  ${hint}`);
-
-    const visible = visibleOptions();
-    for (let i = 0; i < visible.length; i++) {
-      const globalIndex = scrollOffset + i;
-      const opt = optionAt(visible, i);
-      const isCurrent = globalIndex === cursor;
-      const isSelected = selected.has(globalIndex);
-      const prefix = isCurrent ? '\u276f' : ' ';
-      const check = isSelected ? '\u25c9' : '\u25cb';
-      const desc = opt.description
-        ? styledFn(ctx.semantic('muted'), ` \u2014 ${opt.description}`)
-        : '';
-      if (isCurrent && !noColor) {
-        ctx.io.write(`\x1b[K  ${styledFn(ctx.semantic('info'), prefix)} ${styledFn(ctx.semantic('info'), check)} ${boldFn(opt.label)}${desc}\n`);
-      } else if (isSelected && !noColor) {
-        ctx.io.write(`\x1b[K  ${prefix} ${styledFn(ctx.status('success'), check)} ${opt.label}${desc}\n`);
-      } else {
-        ctx.io.write(`\x1b[K  ${prefix} ${check} ${opt.label}${desc}\n`);
-      }
-    }
-  }
-
-  function clearRender(): void {
-    const totalLines = renderLineCount();
-    term.moveUp(totalLines);
-  }
-
-  function cleanup(): void {
-    clearRender();
-    const totalLines = renderLineCount();
-    term.clearBlock(totalLines);
-    const selectedLabels = [...selected].sort().map((i) => optionAt(options.options, i).label).join(', ');
-    const label = formatFormTitle(options.title, ctx) + ' ' + styledFn(ctx.semantic('info'), selectedLabels);
-    ctx.io.write(`\x1b[K${label}\n`);
-    term.showCursor();
-  }
-
-  render();
-
-  return new Promise<T[]>((resolve) => {
-    const handle = subscribeFormKeyInput(ctx, (key) => {
-      const next = handleVerticalNav(key, cursor, options.options.length);
-      if (next !== null) {
-        cursor = next;
-        scrollOffset = clampScroll(cursor, scrollOffset, maxVisible, options.options.length);
-        clearRender(); render();
-      } else if (isKey(key, 'space')) {
-        if (selected.has(cursor)) selected.delete(cursor); else selected.add(cursor);
-        clearRender(); render();
-      } else if (isKey(key, 'enter')) {
-        handle.dispose(); cleanup();
-        resolve([...selected].sort().map((i) => optionAt(options.options, i).value));
-      } else if (isKey(key, 'c', { ctrl: true }) || isKey(key, 'escape')) {
-        selected.clear();
-        handle.dispose(); cleanup(); resolve([]);
-      }
-    });
-  });
 }
